hotfix: handling case where Rex maturities amounts are returned as strings
